+++ Kanikama/Packages/net.shivaduke28.kanikama.core/Editor/ExtensionMethods.cs
@@ -1,0 +1,21 @@
+﻿using System.Linq;
+using UnityEditor;
+using UnityEngine;
+
+namespace Kanikama.Core.Editor
+{
+    public static class ExtensionMethods
+    {
+        public static bool IsContributeGI(this GameObject gameObject)
+        {
+            var flags = GameObjectUtility.GetStaticEditorFlags(gameObject);
+            return flags.HasFlag(StaticEditorFlags.ContributeGI);
+        }
+
+        public static bool IsEmissiveAndContributeGI(this Renderer renderer)
+        {
+            return renderer.gameObject.IsContributeGI() &&
+                renderer.sharedMaterials.Any(m => m.IsEmissive());
+        }
+    }
+}
