+++ Kanikama/Packages/net.shivaduke28.kanikama.core/Editor/GUI/TextureUtilityWindow.cs
@@ -1,4 +1,4 @@
-﻿using Kanikama.Core.Editor.Textures;
+﻿using Kanikama.Core.Editor.Util;
 using UnityEditor;
 using UnityEngine;
