--- Kanikama/Packages/net.shivaduke28.kanikama.core/Editor/KanikamaSceneUtility.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: eb39a6986c37da44cadd74c5eb0bb9c7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
