+++ Kanikama/Packages/net.shivaduke28.kanikama.core/Editor/Util/UnityLightmapUtility.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: dd96efc066e7477386e086c5f9ef3130
+timeCreated: 1683426657
