+++ Kanikama/Packages/net.shivaduke28.kanikama.gi.bakery/Baking/Editor/BakeryBakingSetting.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.IO;
-using Kanikama.Core.Editor.Textures;
+using Kanikama.Core.Editor.Util;
 using UnityEditor;
 using UnityEngine;
