+++ Kanikama/Packages/net.shivaduke28.kanikama.gi.bakery/Baking/Editor/BakeryLightmapper.cs
@@ -20,7 +20,7 @@ namespace Kanikama.GI.Bakery.Editor
         public void SetOutputAssetDirPath(string outputAssetDirPath)
         {
             ftRenderLightmap.useScenePath = false;
-            ftRenderLightmap.outputPath = KanikamaSceneUtility.RemoveAssetsFromPath(outputAssetDirPath);
+            ftRenderLightmap.outputPath = IOUtility.RemoveAssetsFromPath(outputAssetDirPath);
         }
 
         public bool UseMipmap => projectSetting.mipmapLightmaps;
