+++ Kanikama/Packages/net.shivaduke28.kanikama.gi/Baking/Editor/UnityBakingPipelineRunner.cs
@@ -3,6 +3,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Kanikama.Core;
 using Kanikama.Core.Editor;
+using Kanikama.Core.Editor.Util;
 using Kanikama.GI.Baking;
 using UnityEngine;
