+++ Kanikama/Packages/net.shivaduke28.kanikama.gi/Baking/Editor/UnityBakingSetting.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.IO;
 using Kanikama.Core.Editor;
-using Kanikama.Core.Editor.Textures;
+using Kanikama.Core.Editor.Util;
 using UnityEditor;
 using UnityEngine;
